backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pdfplumber>=0.10.3
python-docx>=1.1.0
spacy>=3.7.2
scikit-learn>=1.3.2
joblib>=1.3.2
pydantic>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
requests>=2.31.0
numpy>=1.26.0